MC-4 Scheduled Alarm Entity creation and DB Schema changes

+++ app/src/main/java/com/geekydroid/materialclock/application/db/dao/AlarmDao.kt
@@ -20,5 +20,5 @@ interface AlarmDao {
     fun getAllAlarms(): Flow<List<AlarmMaster>>
 
     @Query("DELETE FROM MC_ALARM_MASTER WHERE alarm_id = :alarmId")
-    suspend fun deleteAlarmById(alarmId: Int)
+    suspend fun deleteAlarmById(alarmId: Long)
 }
